# resource-server/src/main/resources/openapi_arqueo.yml
openapi: 3.0.0
info:
  title: TPV Arqueo API
  version: 1.0.0
paths:
  /arqueo/buscar:
    post:
      summary: Buscar mesas servidas según filtros patron Criteria
      description: Devuelve listado mesas  utilizando patron criteria.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filters:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                      operator:
                        type: string
                      value:
                        type: string
                sortField:
                  type: string
                sortDirection:
                  type: string
                limit:
                  type: integer
                offset:
                  type: integer
                imprimir:
                  type: boolean
                  description: Indica si se debe imprimir un resumen de los resultados de la búsqueda.
                  default: false
      responses:
        '200':
          description: Mesa solicitada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesaServida'
        '400':
          description: Solicitud inválida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Parámetro 'id' es requerido."
  /arqueo/camareros/buscar:
    get:
      summary: Obtener listado único de camareros
      description: Devuelve un listado con los nombres únicos de camareros desde la tabla `mesa_servida`.
      tags:
        - Camareros
      responses:
        '200':
          description: Lista de nombres de camareros
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '500':
          description: Error del servidor    
  /arqueo/arquear:
    post:
      summary: Marcar mesas como arqueadas
      description: Recibe una lista de IDs de mesas y las marca como arqueadas en el sistema.
      tags:
        - Arqueo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                usuario:
                  type: string
              required:
                - ids
      responses:
        '200':
          description: Mesas arqueadas correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Solicitud mal formada
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor
      security:
        - bearerAuth: [ ]
components:
  schemas:
    MesaServida:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mesaReferencia:
          type: string
          format: uuid
        numero:
          type: integer
        sector:
          type: string
        nombre:
          type: string
        ocupada:
          type: boolean
        camarero:
          type: string
        cantidad:
          type: number
          format: double
        fechaInicio:
          type: string
          format: date-time
        fechaFin:
          type: string
          format: date-time
        arqueada:
          type: boolean
        estado:
          type: string
        metodoPago:
          type: string
        borrada:
          type: boolean
        lastUpdatedAt:
          type: string
          format: date-time
        lastUpdatedBy:
          type: string